move autoTypes to Examples
