+++ UI/src/main/kotlin/ca/warp7/rt/core/model/ContextPipeline.kt
@@ -3,10 +3,5 @@ package ca.warp7.rt.core.model
 
 interface ContextPipeline {
     fun fetch()
-    val virtualContext: Context
-    fun streamOf(metrics: Set<Metric<*>>): Stream
-
-    interface Stream {
-        fun mapCols(vararg columns: Pair<String, (PipelineExpression) -> Any?>)
-    }
+    fun streamOf(metrics: Set<Metric<*>>): VectorizedStream
 }
