--- UI/src/main/kotlin/ca/warp7/rt/core/model/Entry.kt
@@ -1,8 +1,0 @@
-package ca.warp7.rt.core.model
-
-class Entry {
-    companion object {
-        val metricsSet: Set<Metric<*>> =
-                teamNumber_ / scout_ / driverStation_ / matchNumber_ / compLevel_ / event_ / year_ / dataSource_
-    }
-}
