+++ UI/src/main/kotlin/ca/warp7/rt/core/model/VectorizedStream.kt
@@ -1,0 +1,5 @@
+package ca.warp7.rt.core.model
+
+interface VectorizedStream {
+    fun mapCols(vararg columns: Pair<String, (PipelineExpression) -> Any?>)
+}
